+++ examples/nuxt-openai/nuxt.config.ts
@@ -1,0 +1,8 @@
+// https://nuxt.com/docs/api/configuration/nuxt-config
+export default defineNuxtConfig({
+  devtools: { enabled: true },
+  modules: ['@nuxtjs/tailwindcss'],
+  runtimeConfig: {
+    openaiApiKey: ''
+  }
+})